feat: added pop and push parser

+++ lib/armv6-m/src/structure.rs
@@ -1,6 +1,6 @@
 /* === Memory === */
 /// Memory is addressed with 32 bits addresses
-pub const MEMORY_MAX_ADDRESSABLE_ADDRESS: u32 = u32::MAX;
+pub const MEMORY_MAX_ADDRESSABLE_ADDRESS: usize = u32::MAX as usize;
 
 /* === Registers === */
